% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dove2.R
\name{dove2}
\alias{dove2}
\title{Estimation of the Vaccine Efficacy}
\usage{
dove2(formula, data, changePts = c(30L, 60L), plots = TRUE)
}
\arguments{
\item{formula}{A formula object, with the response on the left hand side of a
'~' operator, and the covariates and vaccine() function on the right.  
The response must be a survival object as returned by the 'Surv'
function of the \pkg{survival} package. See Details for further information.
The vaccine() function must be used to specify the entry time and
vaccination information. See ?vaccine and Details for further
information.}

\item{data}{A data.frame object. The data.frame in which to interpret the
variable names in formula. Must contain the entry time, the event time,
the event status, the vaccination status, the vaccination time, and the covariates. 
See Details.}

\item{changePts}{A numerical vector object. The potential change
points (in days) of the piece-wise log-linear hazard ratio. The default
change points are 30 and 60 days.}

\item{plots}{A logical object. If TRUE (default), plots of the estimated 
VE_CI (based on KM estimator and the new Cox model), 
the estimated VE_HR and the 95\% confidence interval 
(based on the new Cox model) will be automatically generated 
and saved to a PDF file named ``figure.pdf'' in the current working directory. 
If FALSE, plots will not be generated.}
}
\value{
A list containing four elements, one for each method. Specifically, 
  \itemize{
    \item{KM method: }{A list containing two elements: `Sur_km' and `VE_CI_km'. 
      `Sur_km' is a list containing the time points and survival probabilities of the KM estimates
      for both the placebo and the vaccine groups.
      `VE_CI_km' is a data.frame containing the estimated VE_CI at all jump points, 
      with all events within the first 4 weeks after the 1st dose excluded.}
    \item{Standard Cox model: }{A vector of the estimated vaccine efficacy and its standard error.}
    \item{Standard Poisson model: }{A scalar of the estimated vaccine efficacy.}
    \item{New Cox model: }{A list containing 10 data frames. Four of them contain the estimated
       cumulative incidence for individuals who were randomized on the following four dates:
       (1) first entry time; (2) medium entry time; (3) last entry time; (4) last entry time + 2 months. 
       Four of them contain the estimated VE_CI for individuals who were vaccinated at the above dates, 
       excluding the first 4 weeks of events. One of them contains the estimated VE_HR and the 95\% 
       confidence interval. 
       The last data.frame contains the estimation results on the hazard ratio of each covariate
       (NA if there's no covariate in the model).}
  }
}
\description{
Estimate vaccine efficacy (VE) with Kaplan-Meier (KM) estimator, standard Cox and Poisson models, 
and the new Cox model. 
Event times are potentially right-censored. 
Curves of the estimated VE in reducing the cumulative incidence (VE_CI) 
and the estimated VE in reducing the hazard rate (VE_HR) can be plotted, 
based on the KM analysis or the new Cox model.
}
\details{
The information required for an analysis is 
  \describe{
    \item{Entry Time:}{Calendar time when the participant entered the trial (in days).}
    \item{Event Time:}{Calendar time when the 
       participant experienced symptomatic COVID-19 or when the follow-up ended, 
       whichever occurred first (in days).}
    \item{Event Status:}{Binary indicator on whether symptomatic COVID-19 occurred 
       before the end of follow-up.}
    \item{Vaccination Status:}{Binary indicator taking value 1 if
      vaccination occurred before the end of follow-up and 0 otherwise.}
    \item{Vaccination Time:}{Calendar time when vaccination took place,
      with an arbitrary value if the participant was not vaccinated.}
    \item{Covariates:}{Baseline covariates (e.g., priority group, age, sex, ethnicity).}
   }
   
Note that all the time variables are measured from the beginning of the 
clinical trial and are specified in days. For each 
individual, the entry_time and event_time must satisfy
entry_time \eqn{\le} event_time. If entry_time > event_time, the case will be 
removed from the analysis and a message will be generated.  

The general structure of the formula input is
  \preformatted{
  Surv(event_time, event_status) ~ covariates + 
    vaccine(entry_time, vaccination_status, vaccination_time)
  }

The response variable must be a survival object as returned by the
'Surv()' function of package \pkg{survival}, where event_time is the
follow up time (formal argument 'time') and event_status is the status
indicator input (formal argument 'event'). Specifically, 
\preformatted{Surv(time = event_time, event = event_status)}

The covariates can be either numerical or categorical.
A model without covariates is also allowed.

The vaccination and entry_time information must be specified through function 
'vaccine()'. Specifically, 
\preformatted{vaccine(entry_time, vaccination_status, vaccination_time)}
}
\examples{
data(dove2Data)

set.seed(1234)
smp <- sample(1L:nrow(x = dove2Data), size = 500L)

# Include priority and sex as covariates
res <- dove2(formula = Surv(event.time, event.status) ~ priority + sex + 
                       vaccine(entry.time, vaccine.status, vaccine.time),
             data = dove2Data[smp,])

}
