% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\name{vaccine}
\alias{vaccine}
\title{Specify Vaccination Variables and Entry Time}
\usage{
vaccine(entry_time, vaccination_status, vaccination_time)
}
\arguments{
\item{entry_time}{The variable for the time when 
the participant entered the trial.}

\item{vaccination_status}{The variable indicating the vaccination 
status: 1 = vaccinated; 0 = not vaccinated.}

\item{vaccination_time}{The variable for the time when 
vaccination took place, with an arbitrary value if the participant was not vaccinated.}
}
\value{
This function is intended to be used only in the model statement 
 of dove2(). The result, a matrix, is used internally.
}
\description{
This function is used in the model statement of dove2() to specify
 the vaccination time, vaccination status, and entry time.
}
